--- api/bonsai_api/old_models/__init__.py
@@ -1,1 +1,0 @@
-"""Collection of data models used for validating API input data and for database schema."""
