+++ base-image/bonsai_models/db/__init__.py
@@ -1,0 +1,1 @@
+"""Data stored in database."""
