+++ base-image/bonsai_models/db/group.py
@@ -1,0 +1,7 @@
+"""Database representation of group information."""
+
+from bonsai_models.schema.group import GroupBase
+
+
+class GroupInDb(GroupBase):  # pylint: disable=too-few-public-methods
+    """Defines group info stored in the databas."""
