+++ base-image/bonsai_models/schema/__init__.py
@@ -1,0 +1,1 @@
+"""API input/ output validation."""
