+++ base-image/bonsai_models/schema/metadata.py
@@ -1,20 +1,10 @@
 """Metadata models."""
 
 from datetime import datetime
-from enum import StrEnum
 from typing import Literal
-
 from pydantic import BaseModel
 
-from .base import RWModel
-
-
-class MetadataTypes(StrEnum):
-    """Valid datatypes for metadata records."""
-
-    STR = "string"
-    INT = "integer"
-    FLOAT = "float"
+from bonsai_models.constants import MetadataTypes
 
 
 class GenericMetadataEntry(BaseModel):
@@ -58,44 +48,3 @@ class TableMetadataInDb(BaseModel):
 InputMetaEntry = DatetimeMetadataEntry | InputTableMetadata | GenericMetadataEntry
 MetaEntryInDb = DatetimeMetadataEntry | TableMetadataInDb | GenericMetadataEntry
 MetaEntriesInDb = list[MetaEntryInDb]
-
-
-class SoupType(StrEnum):
-    """Type of software of unkown provenance."""
-
-    DB = "database"
-    SW = "software"
-
-
-class SoupVersion(BaseModel):
-    """Version of Software of Unknown Provenance."""
-
-    name: str
-    version: str
-    type: SoupType
-
-
-class SequencingInfo(RWModel):
-    """Information on the sample was sequenced."""
-
-    run_id: str
-    platform: str
-    instrument: str | None = None
-    method: dict[str, str] = {}
-    date: datetime | None = None
-
-
-class PipelineInfo(RWModel):
-    """Information on the sample was analysed."""
-
-    pipeline: str
-    version: str
-    commit: str
-    analysis_profile: list[str]
-    assay: str
-    release_life_cycle: str
-    configuration_files: list[str]
-    workflow_name: str
-    command: str
-    softwares: list[SoupVersion]
-    date: datetime
