+++ base-image/bonsai_models/schema/pipeline/__init__.py
@@ -1,0 +1,1 @@
+"""Definition of input data."""
