+++ base-image/bonsai_models/schema/pipeline/metadata.py
@@ -1,0 +1,41 @@
+"""Metadata from the pipeline."""
+
+from datetime import datetime
+from pydantic import BaseModel
+from bonsai_models.base import RWModel
+
+from .constants import SoupType
+
+
+class SoupVersion(BaseModel):
+    """Version of Software of Unknown Provenance."""
+
+    name: str
+    version: str
+    type: SoupType
+
+
+class SequencingInfo(RWModel):
+    """Information on the sample was sequenced."""
+
+    run_id: str
+    platform: str
+    instrument: str | None = None
+    method: dict[str, str] = {}
+    date: datetime | None = None
+
+
+class PipelineInfo(RWModel):
+    """Information on the sample was analysed."""
+
+    pipeline: str
+    version: str
+    commit: str
+    analysis_profile: list[str]
+    assay: str
+    release_life_cycle: str
+    configuration_files: list[str]
+    workflow_name: str
+    command: str
+    softwares: list[SoupVersion]
+    date: datetime
