--- base-image/bonsai_models/util.py
@@ -1,8 +1,0 @@
-
-import datetime
-
-
-def get_timestamp() -> datetime.datetime:
-    """Get datetime timestamp in utc timezone."""
-    return datetime.datetime.now(tz=datetime.UTC)
-
