+++ frontend/bonsai_app/models.py
@@ -10,7 +10,6 @@ class RWModel(BaseModel):  # pylint: disable=too-few-public-methods
     """Base model for read/ write operations"""
 
     model_config = ConfigDict(
-        allow_population_by_alias=True,
         populate_by_name=True,
         use_enum_values=True,
     )
